LPS-119068 Create intermediate component to setup listeners al SideNavigation
